requirements.txt
# 1. WEB FRAMEWORK & CORE 
Flask==3.1.2               
Werkzeug==3.1.5             
Jinja2==3.1.6              
MarkupSafe==3.0.3           
itsdangerous==2.2.0         
click==8.3.1                
blinker==1.9.0              

# 2. DATABASE & ORM 
SQLAlchemy==2.0.46         
Flask-SQLAlchemy==3.1.1     
pymongo==4.16.0             
greenlet==3.3.1             
dnspython==2.8.0           

# 3. SECURITY & CONFIG
flask-talisman==1.1.0      
flask-cors==6.0.2           
python-dotenv==1.2.1        
PySocks==1.7.1              

# 4. RUNTIME & PRODUCTION
gunicorn==25.1.0            
honcho==2.0.0               

# 5. TESTING & DATA MOCKING
pytest==9.0.2               
factory_boy==3.3.3          
Faker==40.4.0               
coverage==7.13.4            
pluggy==1.6.0               
iniconfig==2.3.0            

# 6. LINTING & FORMATTING
pylint==4.0.4               
astroid==4.0.4              
flake8==7.3.0               
pycodestyle==2.14.0         
pyflakes==3.4.0             
black==26.1.0               
isort==7.0.0               
mccabe==0.7.0               

# 7. UTILITIES & TOOLS
httpie==3.2.4               
Pygments==2.19.2            
markdown-it-py==4.0.0       
mdurl==0.1.2                
